pyproject.toml
[project]
name = "school-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-adk>=1.12.0",
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "python-dotenv>=1.1.1",
    "pyyml>=0.0.2",
    "uvicorn[standard]>=0.35.0",
]
